fix: OnEvent is not propagating data in some scenarios with postgres (#637)
* chore: simplify stream_pipeline_request

* fix: multiple client subscribe

* chore: add log if broadcast error

* fix: Fix `Field` to `grpc::types::Value` conversion

* chore: fix issue delay async

+++ dozer-api/src/grpc/types_helper.rs
@@ -47,7 +47,7 @@ fn field_to_prost_value(f: Field) -> Value {
             value: Some(value::Value::IntValue(n)),
         },
         Field::Float(n) => Value {
-            value: Some(value::Value::FloatValue(n.0 as f32)),
+            value: Some(value::Value::DoubleValue(n.0)),
         },
 
         Field::Boolean(n) => Value {
